# Kramp.API/Swagger/V1.3.0_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kramp API",
    "description": "Kramp API",
    "contact": {
      "name": "Página de Contato",
      "url": "https://www.exemplo.com.br"
    },
    "license": {
      "name": "Licença",
      "url": "https://www.exemplo.com.br"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Workout/Create": {
      "post": {
        "tags": [
          "Workout"
        ],
        "summary": "Cria um novo treino (Workout).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutCommand"
              },
              "examples": {
                "requisicao-valida": {
                  "summary": "Exemplo de requisição válida",
                  "value": {
                    "name": "Treino de Peito e Tríceps",
                    "description": "Foco em hipertrofia para peito e tríceps.",
                    "period": 2,
                    "exercises": [
                      {
                        "exerciseId": "565c4a54-e4f8-4c1c-a3be-08db428e781e",
                        "workoutId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "restTimeInSeconds": 60,
                        "exerciseTimeInSeconds": 30,
                        "series": 3,
                        "repetitions": 12
                      },
                      {
                        "exerciseId": "a1b2c3d4-e5f6-4d1e-a4bf-09ec539f892f",
                        "workoutId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "restTimeInSeconds": 45,
                        "exerciseTimeInSeconds": 45,
                        "series": 4,
                        "repetitions": 10
                      }
                    ]
                  }
                },
                "requisicao-invalida": {
                  "summary": "Exemplo de requisição inválida (faltando nome)",
                  "value": {
                    "description": "Foco em hipertrofia para peito e tríceps.",
                    "period": 2,
                    "exercises": [
                      {
                        "exerciseId": "565c4a54-e4f8-4c1c-a3be-08db428e781e",
                        "workoutId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "restTimeInSeconds": 60,
                        "exerciseTimeInSeconds": 30,
                        "series": 3,
                        "repetitions": 12
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                },
                "examples": {
                  "resposta-201": {
                    "summary": "Exemplo de resposta (201 Created)",
                    "value": {
                      "name": "Treino de Peito e Tríceps",
                      "description": "Foco em hipertrofia para peito e tríceps.",
                      "seriesCount": 0,
                      "repetitionCount": 0,
                      "period": 2,
                      "targetedMuscles": [],
                      "exercises": [],
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "createdAt": "2023-12-06T12:34:56Z",
                      "updatedAt": "2023-12-06T12:34:56Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "nome-obrigatorio": {
                    "summary": "Erro - Nome é obrigatório",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Um ou mais erros de validação ocorreram.",
                      "status": 400,
                      "errors": {
                        "Name": [
                          "O nome é obrigatório."
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "erro-inesperado": {
                    "summary": "Erro interno durante a criação do treino.",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Erro interno do servidor.",
                      "status": 500,
                      "detail": "Ocorreu um erro inesperado ao processar a solicitação."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Workout/All": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Retorna todos os treinos (Workouts).",
        "description": "Este endpoint recupera uma lista de todos os treinos cadastrados no sistema.",
        "responses": {
          "200": {
            "description": "Sucesso - Lista de treinos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "Nome do treino.",
                        "maxLength": 50,
                        "minLength": 3,
                        "nullable": true
                      },
                      "description": {
                        "type": "string",
                        "description": "Descrição detalhada do treino.",
                        "maxLength": 240,
                        "nullable": true
                      },
                      "seriesCount": {
                        "type": "integer",
                        "description": "Número total de séries do treino."
                      },
                      "repetitionCount": {
                        "type": "integer",
                        "description": "Número de repetições por série no treino."
                      },
                      "period": {
                        "type": "integer",
                        "description": "Define a frequência de repetição do treino, considerando os dias de descanso.\n",
                        "enum": [
                          1,
                          2,
                          3,
                          4,
                          5
                        ],
                        "example": 1,
                        "x-enum-descriptions": {
                          "1": "DAILY",
                          "2": "WEEKLY",
                          "3": "JUMP_ONE_DAY",
                          "4": "JUMP_TWO_DAYS",
                          "5": "JUMP_THREE_DAYS"
                        }
                      },
                      "targetedMuscles": {
                        "type": "array",
                        "description": "Lista de músculos alvo do treino.",
                        "items": {
                          "type": "integer",
                          "description": "Representa um grupo muscular alvo de exercícios.",
                          "enum": [
                            1,
                            2,
                            3,
                            4,
                            5,
                            6,
                            7,
                            8,
                            9,
                            10,
                            11
                          ],
                          "x-enum-descriptions": {
                            "1": "PEITO",
                            "2": "BÍCEPS",
                            "3": "TRÍCEPS",
                            "4": "COSTAS",
                            "5": "OMBRO",
                            "6": "ANTEBRAÇOS",
                            "7": "ABDÔMEN",
                            "8": "QUADRÍCEPS",
                            "9": "ISQUIOTIBIAIS",
                            "10": "GLÚTEOS",
                            "11": "PANTURRILHAS"
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "sucesso": {
                    "summary": "Lista de treinos.",
                    "value": [
                      {
                        "name": "Treino de Peito e Tríceps",
                        "description": "Foco em hipertrofia para peito e tríceps.",
                        "seriesCount": 3,
                        "repetitionCount": 12,
                        "period": 2,
                        "targetedMuscles": [
                          "CHEST",
                          "TRICEPS"
                        ]
                      },
                      {
                        "name": "Treino de Costas e Bíceps",
                        "description": "Intenso treino para costas e bíceps.",
                        "seriesCount": 4,
                        "repetitionCount": 10,
                        "period": 1,
                        "targetedMuscles": [
                          "BACK",
                          "BICEPS"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Interno no Servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "erro-inesperado": {
                    "summary": "Erro interno ao recuperar treinos.",
                    "value": {
                      "error": "Ocorreu um erro inesperado ao processar sua solicitação. Por favor, tente novamente mais tarde."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Workout/{Id}": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Retorna um treino específico pelo ID.",
        "description": "Este endpoint recupera os detalhes de um treino específico utilizando o ID fornecido.",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "O UUID do treino que deseja recuperar.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Informações do treino retornadas com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutInfoViewModel"
                },
                "examples": {
                  "successResponse": {
                    "summary": "Exemplo de resposta de sucesso",
                    "value": {
                      "name": "Treino A",
                      "description": "Treino para iniciantes",
                      "seriesCount": 3,
                      "repetitionCount": 12,
                      "period": "DAILY",
                      "targetedMuscles": [
                        "CHEST",
                        "TRICEPS"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Treino não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Treino não encontrado com o ID fornecido."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Workout/Update/{Id}": {
      "put": {
        "tags": [
          "Workout"
        ],
        "summary": "Atualiza um treino específico.",
        "description": "Este endpoint permite a atualização das informações de um treino utilizando o ID fornecido.",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "O UUID do treino que deseja atualizar.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutCommand"
              },
              "examples": {
                "validRequest": {
                  "summary": "Exemplo de uma requisição válida",
                  "value": {
                    "Name": "Treino A - Atualizado",
                    "Description": "Treino para intermediários",
                    "Period": "WEEKLY",
                    "Exercises": []
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Treino atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                },
                "examples": {
                  "successResponse": {
                    "summary": "Exemplo de resposta de sucesso",
                    "value": {
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "name": "Treino A - Atualizado",
                      "description": "Treino para intermediários",
                      "createdAt": "2023-10-26T12:34:56Z",
                      "updatedAt": "2023-10-26T13:34:56Z",
                      "seriesCount": 3,
                      "repetitionCount": 12,
                      "period": "WEEKLY",
                      "targetedMuscles": [
                        "CHEST",
                        "TRICEPS"
                      ],
                      "exercises": []
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Treino não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Treino não encontrado com o ID fornecido."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Workout/Delete/{Id}": {
      "delete": {
        "tags": [
          "Workout"
        ],
        "summary": "Exclui um treino específico.",
        "description": "Este endpoint permite a exclusão de um treino específico utilizando o ID fornecido.",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "O UUID do treino que deseja excluir.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content - Treino excluído com sucesso."
          },
          "404": {
            "description": "Treino não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Treino não encontrado com o ID fornecido."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateWorkoutCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do treino (opcional)",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Descrição do treino (opcional)",
            "nullable": true
          },
          "period": {
            "$ref": "#/components/schemas/Period",
            "description": "Define a frequência de repetição do treino."
          },
          "exercises": {
            "type": "array",
            "description": "Lista de exercícios do treino (opcional)",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/WorkoutExercise"
            }
          }
        }
      },
...
      "UpdateWorkoutCommand": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true,
            "description": "Identificador único do treino. Este campo é ignorado na requisição e definido internamente.",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "Name": {
            "type": "string",
            "nullable": true,
            "description": "Nome do treino.",
            "example": "Treino A - Atualizado"
          },
          "Description": {
            "type": "string",
            "nullable": true,
            "description": "Descrição do treino.",
            "example": "Treino para intermediários"
          },
          "Period": {
            "$ref": "#/components/schemas/Period",
            "description": "Define a frequência de repetição do treino."
          },
          "Exercises": {
            "type": "array",
            "description": "Lista de exercícios do treino.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/WorkoutExercise"
            }
          }
        }
      },
...
      "WorkoutInfoViewModel": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string",
            "nullable": true,
            "description": "Nome do treino."
          },
          "Description": {
            "type": "string",
            "nullable": true,
            "description": "Descrição do treino."
          },
          "SeriesCount": {
            "type": "integer",
            "description": "Número total de séries do treino."
          },
          "RepetitionCount": {
            "type": "integer",
            "description": "Número de repetições por série no treino."
          },
          "Period": {
            "$ref": "#/components/schemas/Period",
            "description": "Define a frequência de repetição do treino (diário, semanal, etc.).",
            "nullable": true
          },
          "TargetedMuscles": {
            "type": "array",
            "description": "Lista de músculos alvo do treino.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Muscle"
            }
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Um URI de referência sobre o tipo de problema."
          },
          "title": {
            "type": "string",
            "description": "Um título curto que descreve o problema."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "O código de status HTTP correspondente ao problema."
          },
          "detail": {
            "type": "string",
            "description": "Uma explicação detalhada do problema, geralmente incluindo informações da exceção."
          },
          "instance": {
            "type": "string",
            "description": "Um URI que identifica a instância específica do problema."
          },
          "errorCode": {
            "type": "string",
            "description": "Um código de erro específico que indica o tipo de falha da aplicação."
          },
          "traceId": {
            "type": "string",
            "description": "Um identificador único para rastrear a requisição específica no sistema."
          },
          "data": {
            "type": "object",
            "description": "Dados adicionais fornecidos pela exceção, geralmente valores chave do dicionário de dados da exceção."
          }
        },
        "required": [
          "title",
          "status"
        ],
        "example": {
          "type": "https://example.com/probs/out-of-credit",
          "title": "Erro interno no servidor.",
          "status": 500,
          "detail": "System.NullReferenceException = Object reference not set to an instance of an object.",
          "instance": "/minha-api/v1/resource",
          "errorCode": "500_ERR",
          "traceId": "00-ab123cd4ef567gh8-1234567890abcdef-01",
          "data": {
            "ChaveExemplo": "ValorExemplo"
          }
        }
      },
      "WorkoutExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true,
            "description": "Identificador único do WorkoutExercise.  Este campo é ignorado na requisição e definido internamente.\n"
          },
          "exerciseId": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador do Exercício (Exercise) relacionado."
          },
          "workoutId": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador do Treino (Workout) relacionado."
          },
          "restTimeInSeconds": {
            "type": "integer",
            "description": "Tempo de descanso entre as séries em segundos."
          },
          "exerciseTimeInSeconds": {
            "type": "integer",
            "description": "Tempo de duração do exercício em segundos."
          },
          "series": {
            "type": "integer",
            "description": "Número de séries do exercício."
          },
          "repetitions": {
            "type": "integer",
            "description": "Número de repetições por série."
          }
        },
        "required": [
          "exerciseId",
          "workoutId",
          "restTimeInSeconds",
          "series",
          "repetitions"
        ]
      },
      "Workout": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true,
            "description": "Identificador único do treino. Este campo é gerado automaticamente e não deve ser fornecido na requisição."
          },
          "name": {
            "type": "string",
            "description": "Nome do treino.",
            "maxLength": 50,
            "minLength": 3,
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Descrição detalhada do treino.",
            "maxLength": 240,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data e hora de criação do treino.",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data e hora da última atualização do treino.",
            "readOnly": true
          },
          "seriesCount": {
            "type": "integer",
            "description": "Número total de séries do treino."
          },
          "repetitionCount": {
            "type": "integer",
            "description": "Número de repetições por série no treino."
          },
          "period": {
            "$ref": "#/components/schemas/Period",
            "description": "Define a frequência de repetição do treino."
          },
          "targetedMuscles": {
            "type": "array",
            "description": "Lista de músculos alvo do treino.",
            "items": {
              "$ref": "#/components/schemas/Muscle"
            }
          },
          "exercises": {
            "type": "array",
            "description": "Lista de exercícios que compõem o treino.",
            "items": {
              "$ref": "#/components/schemas/WorkoutExercise"
            }
          }
        }
      },
...
      "Period": {
        "type": "integer",
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "description": "Define a frequência de repetição do treino, considerando os dias de descanso.\n",
        "example": 1,
        "x-enum-descriptions": {
          "1": "DAILY",
          "2": "WEEKLY",
          "3": "JUMP_ONE_DAY",
          "4": "JUMP_TWO_DAYS",
          "5": "JUMP_THREE_DAYS"
        }
      },
      "Muscle": {
        "type": "integer",
        "description": "Representa um grupo muscular alvo de exercícios.",
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11
        ],
        "x-enum-descriptions": {
          "1": "PEITO",
          "2": "BÍCEPS",
          "3": "TRÍCEPS",
          "4": "COSTAS",
          "5": "OMBRO",
          "6": "ANTEBRAÇOS",
          "7": "ABDÔMEN",
          "8": "QUADRÍCEPS",
          "9": "ISQUIOTIBIAIS",
          "10": "GLÚTEOS",
          "11": "PANTURRILHAS"
        }
      }
    }
  }
}